+++ spec/test_app/app/controllers/columns_controller.rb
@@ -1,0 +1,7 @@
+class ColumnsController < ApplicationController
+  before_filter :get_npw
+
+  def get_npw
+    @npw = NobelPrizeWinner.select('nobel_prize_winners.*, nobel_prizes.category, nobel_prizes.year').joins(:nobel_prizes).tableficate(params[:nobel_prize_winners])
+  end
+end