+++ changelog.markdown
@@ -9,6 +9,7 @@
  * `column` now takes a `header_attrs` option to provide attributes for the `th`
  * `actions` now takes options just like `column`
  * `column` now takes a `cell_attrs` option to provide attributes for every `td` in the column
+ * `column` now places all unrecognized options into the `attrs` attribute which is used on a `col` tag
 
 ## 0.2.1
  * fix table generator to provide `scope` the proper argument and correct documentation